tests/wpt/tests/server-timing/server_timing_headers_not_visible_in_fetch.https.html

<!DOCTYPE html>
<head>
    <meta charset='utf-8' />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
    <script>
      const {HTTPS_REMOTE_ORIGIN} = get_host_info()
      promise_test(async t => {
        const response = await fetch(`${HTTPS_REMOTE_ORIGIN}/server-timing/resources/blue_cors.png`, {mode: "cors"});
        assert_equals(response.headers.has("server-timing"), false);
      }, "Server-Timing headers should be explicitly allowed by CORS");
    </script>
</head>
